requirements.txt
playwright>=1.50.0
beautifulsoup4>=4.12.3
supabase>=2.10.0
python-telegram-bot>=21.7
httpx>=0.28.1
python-dotenv>=1.0.1
